Move components that need access to the state inside of src/components. Leave UI components in root components folder.
