+++ src/app/Components/EthrDid/DelegateContainer.ts
@@ -2,8 +2,8 @@ import { Authentication } from 'did-resolver'
 import { connect } from 'react-redux'
 import { AnyAction } from 'redux'
 import { ThunkDispatch } from 'redux-thunk'
-import { stateInterface } from '../../app/state/configureStore'
-import { addDelegate } from '../../app/state/operations/ethrdid'
+import { stateInterface } from '../../state/configureStore'
+import { addDelegate } from '../../state/operations/ethrdid'
 import DelegateComponent from './DelegateComponent'
 
 const mapStateToProps = (state: stateInterface) => ({
